메인페이지 리팩토링 (리덕스 및 타입스크립트 추가 리팩토링 예정  ...)
